Refactor labor supply response module into individual variable files
- Split labor_supply_response.py into 9 separate variable files
- Add unit tests for each variable following CTC pattern
- Maintain all original functionality and dependencies
- Improve code organization and modularity

This prepares the module for fixing negative self-employment income handling issues.
