+++ policyengine_us/tests/policy/baseline/gov/simulation/labor_supply_response/labor_supply_behavioral_response.yaml
@@ -1,0 +1,21 @@
+- name: No elasticities means no response
+  period: 2023
+  input:
+    employment_income_before_lsr: 50_000
+    self_employment_income_before_lsr: 10_000
+    gov.simulation.labor_supply_responses.elasticities.income: 0
+    gov.simulation.labor_supply_responses.elasticities.substitution.all: 0
+  output:
+    labor_supply_behavioral_response: 0
+
+- name: Zero income and substitution elasticities
+  period: 2023
+  input:
+    employment_income_before_lsr: 50_000
+    self_employment_income_before_lsr: 10_000
+    income_elasticity_lsr: 0
+    substitution_elasticity_lsr: 0
+    gov.simulation.labor_supply_responses.elasticities.income: 0
+    gov.simulation.labor_supply_responses.elasticities.substitution.all: 0
+  output:
+    labor_supply_behavioral_response: 0
