+++ policyengine_us/tests/policy/baseline/gov/simulation/labor_supply_response/relative_income_change.yaml
@@ -1,0 +1,6 @@
+- name: Default bounds parameter
+  period: 2023
+  input:
+    gov.simulation.labor_supply_responses.bounds.income_change: 1.0
+  output:
+    relative_income_change: 0
