+++ policyengine_us/tests/policy/baseline/gov/simulation/labor_supply_response/self_employment_income_behavioral_response.yaml
@@ -1,0 +1,31 @@
+- name: No labor supply response
+  period: 2023
+  input:
+    labor_supply_behavioral_response: 0
+    employment_income_behavioral_response: 0
+  output:
+    self_employment_income_behavioral_response: 0
+
+- name: All employment income response
+  period: 2023
+  input:
+    labor_supply_behavioral_response: 1_000
+    employment_income_behavioral_response: 1_000
+  output:
+    self_employment_income_behavioral_response: 0
+
+- name: Mixed income response
+  period: 2023
+  input:
+    labor_supply_behavioral_response: 1_200
+    employment_income_behavioral_response: 800
+  output:
+    self_employment_income_behavioral_response: 400
+
+- name: Negative self-employment response
+  period: 2023
+  input:
+    labor_supply_behavioral_response: 1_000
+    employment_income_behavioral_response: 1_250
+  output:
+    self_employment_income_behavioral_response: -250
