+++ policyengine_us/tests/policy/baseline/gov/simulation/labor_supply_response/substitution_elasticity_lsr.yaml
@@ -1,0 +1,9 @@
+- name: Zero substitution elasticity gives no response
+  period: 2023
+  input:
+    employment_income_before_lsr: 50_000
+    self_employment_income_before_lsr: 10_000
+    substitution_elasticity: 0
+    gov.simulation.labor_supply_responses.elasticities.substitution.all: 0
+  output:
+    substitution_elasticity_lsr: 0
